Referencing finmath-lib 3.2.10

+++ src/net/finmath/experiments/montecarlo/assetderivativevaluation/products/EuropeanOptionGammaLikelihood.java
@@ -74,8 +74,8 @@ public class EuropeanOptionGammaLikelihood extends AbstractAssetMonteCarloProduc
 				// Get some model parameters
 				double T		= maturity;
 				double S0		= underlyingAtToday.get(path);
-				double r		= blackScholesModel.getModel().getRiskFreeRate();
-				double sigma	= blackScholesModel.getModel().getVolatility();
+				double r		= blackScholesModel.getModel().getRiskFreeRate().doubleValue();
+				double sigma	= blackScholesModel.getModel().getVolatility().doubleValue();
 
 				double ST		= underlyingAtMaturity.get(path);
